break down into smaller files and refactor some to typescript
