+++ packages/mongo/oplog_observe_driver.js
@@ -2,6 +2,11 @@ import has from 'lodash.has';
 import isEmpty from 'lodash.isempty';
 import { oplogV2V1Converter } from "./oplog_v2_converter";
 import { check, Match } from 'meteor/check';
+import { CursorDescription } from './cursor_description';
+import { forEachTrigger, listenAll } from './mongo_driver';
+import { Cursor } from './cursor';
+import LocalCollection from 'meteor/minimongo/local_collection';
+import { idForOp } from './oplog_tailing';
 
 var PHASE = {
   QUERYING: "QUERYING",
@@ -30,7 +35,7 @@ var currentId = 0;
 // same simple interface: constructing it starts sending observeChanges
 // callbacks (and a ready() invocation) to the ObserveMultiplexer, and you stop
 // it by calling the stop() method.
-OplogObserveDriver = function (options) {
+export const OplogObserveDriver = function (options) {
   const self = this;
   self._usesOplog = true;  // tests look at this
 
@@ -113,9 +118,6 @@ OplogObserveDriver = function (options) {
 
   self._requeryWhenDoneThisQuery = false;
   self._writesToCommitWhenWeReachSteady = [];
-
-
-
  };
 
 _.extend(OplogObserveDriver.prototype, {
@@ -1052,5 +1054,3 @@ var modifierCanBeDirectlyApplied = function (modifier) {
     });
   });
 };
-
-MongoInternals.OplogObserveDriver = OplogObserveDriver;
